requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
requests
django-otp
django-otp-yubikey
